Task/requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
